Fix signedness in Msg

+++ src/Master.cxx
@@ -31,7 +31,7 @@ Master::Master(MainOpt &config) : config(config), command_listener(config) {
 
 void Master::handle_command_message(std::unique_ptr<KafkaW::Msg> &&msg) {
   CommandHandler command_handler(config, this);
-  command_handler.handle({(char *)msg->data(), (int32_t)msg->size()});
+  command_handler.handle({(char *)msg->data(), msg->size()});
 }
 
 void Master::handle_command(rapidjson::Document const &cmd) {
